requirements.txt
PyQt5
python-dotenv
markdown
google-generativeai
PyPDF2
pygraphviz
